Restructure file hierarchy by topic/*.ts and topic/components/*
