+++ arches/app/src/controlledlists/components/editor/ItemURI.vue
@@ -6,16 +6,16 @@ import InputText from "primevue/inputtext";
 import { useToast } from "primevue/usetoast";
 
 import { ARCHES_CHROME_BLUE } from "@/theme.ts";
-import { patchItem } from "@/components/ControlledListManager/api.ts";
+import { patchItem } from "@/controlledlists/api.ts";
 import {
     DEFAULT_ERROR_TOAST_LIFE,
     ERROR,
     itemKey,
-} from "@/components/ControlledListManager/constants.ts";
-import { vFocus } from "@/components/ControlledListManager/utils.ts";
+} from "@/controlledlists/constants.ts";
+import { vFocus } from "@/controlledlists/utils.ts";
 
 import type { Ref } from "vue";
-import type { ControlledListItem } from "@/types/ControlledListManager";
+import type { ControlledListItem } from "@/controlledlists/types";
 
 const item = inject(itemKey) as Ref<ControlledListItem>;
