+++ arches/app/src/plugins/ControlledListManager.vue
@@ -1,14 +1,14 @@
 <script setup lang="ts">
 import Toast from "primevue/toast";
 
-import ControlledListEditor from "@/components/ControlledListManager/ControlledListEditor.vue";
+import ControlledListsMain from "@/controlledlists/components/ControlledListsMain.vue";
 </script>
 
 <template>
     <!-- Subtract size of arches toolbars -->
     <div style="width: calc(100vw - 50px); height: calc(100vh - 50px)">
         <div style="height: 100%">
-            <ControlledListEditor />
+            <ControlledListsMain />
         </div>
     </div>
     <Toast />
